feat(radiation): route radiation damage through RadiationCapability's dose/contagion model
Radiation damage sources (campfire, fan enriching, the RADIATION mob
effect) previously applied the effect/damage directly; they now go
through RadiationCapability's dose-based contagion system, and the
fan_radiation damage type (previously commented out) is wired up and
tagged as NO_KNOCKBACK alongside RADIATION.

- RadiationEffect: extend the new VicinityEffect base instead of
  MobEffect directly, adding a contagion radius/duration and an
  onContaminate(...) hook that calls
  RadiationCapability.applyContagion(...) on nearby entities;
  applyEffectTick() now defers to super, computes damage from
  RadiationCapability.getRadiationResistance(...), and deals it via
  CNDamageSources.radiation(...) instead of livingEntity.damageSources().magic()
- EnrichingCampfireBlock: entityInside() now calls
  RadiationCapability.applyContagion(livingEntity, CAMPFIRE_DOSE, 100)
  instead of directly adding a RADIATION MobEffectInstance
- CNFanProcessingTypes.EnrichedType: affectEntity() now applies a
  contagion dose via RadiationCapability.applyContagion(...) and
  separately deals 1 point of fan_radiation damage via
  CNDamageSources.fanRadiation(level), instead of only adding a
  RADIATION MobEffectInstance
- CNDamageSources: un-comment and enable fanRadiation(Level), which
  was previously dead code
- CNDamageTypes: clean up radiation(Level) to use imported DamageSource/
  Level types instead of fully-qualified names (no behavior change)
- Add CNDamageTypeTagsProvider: a plain TagsProvider<DamageType>
  (not a Registrate dynamic tag generator, since DamageType entries
  only resolve through the datapack-registry HolderLookup.Provider)
  tagging RADIATION and FAN_RADIATION under DamageTypeTags.NO_KNOCKBACK
- CreateNuclearDatagen: register the new CNDamageTypeTagsProvider as
  a server data provider
- CameraAccessor: widen callMove(...)'s parameters from double to
  float, matching the updated Camera#move signature
- GameRendererMixin: update the @Inject target from the old
  render(FJZ)V signature to the current
  render(Lnet/minecraft/client/DeltaTracker;Z)V signature, and read
  the partial tick via deltaTracker.getGameTimeDeltaPartialTick(false)
  instead of a raw float parameter

+++ src/main/java/net/nuclearteam/createnuclear/foundation/mixin/client/CameraAccessor.java
@@ -8,5 +8,5 @@ import org.spongepowered.asm.mixin.gen.Invoker;
 public interface CameraAccessor {
     // Exposes the private Camera#move so it can be called from outside the class
     @Invoker("move")
-    void callMove(double distanceOffset, double verticalOffset, double horizontalOffset);
+    void callMove(float distanceOffset, float verticalOffset, float horizontalOffset);
 }
